+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ## [0.1.0-alpha48] - 2024-10-15
 
 - windows adaptation
+- improve: `tagger tag` will display tags in each image's title
 
 ## [0.1.0-alpha47] - 2024-10-14